Fix zoomLevel Double->Int cast in RadarTileOverlay (Build #10)
OSMDroid 6.1.18 returns Projection.zoomLevel as Double; add .toInt()
before coerceIn so the shl and tile-coordinate calculations get an Int.

+++ MeteoRadar/app/src/main/java/it/shinyup/meteoradar/ui/radar/RadarTileOverlay.kt
@@ -47,7 +47,7 @@ class RadarTileOverlay(
 
         val frame = frames.getOrNull(currentFrameIndex) ?: return
         val projection = mapView.projection
-        val zoom = projection.zoomLevel.coerceIn(0, 18)
+        val zoom = projection.zoomLevel.toInt().coerceIn(0, 18)
         val numTiles = 1 shl zoom
         val bbox = projection.boundingBox
